refactored

+++ RevitRuntimeCompiler/CodeProvider/CSharpCodeProvider.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.IO;
 using System.Reflection;
+using Microsoft.Win32;
 
 namespace RevitRuntimeCompiler.CodeProvider
 {
@@ -15,21 +16,30 @@ namespace RevitRuntimeCompiler.CodeProvider
         private readonly string _copySolutionPath;
         private readonly string _codePath;
         private readonly string _thisDir;
+        private readonly string _solutionAndProjectName = "EditorSolution";
 
-        public string ExecutableFilePath => _copySolutionPath + @"\EditorSolution.sln";
+        public string ExecutableFilePath => _copySolutionPath + @$"\{_solutionAndProjectName}.sln";
 
         public CSharpCodeProvider(string revitVersion)
         {
             _thisDir = Path.GetDirectoryName(Assembly.GetAssembly(GetType()).Location);
-            _originalSolutionPath = _thisDir + @"\EditorSolution";
+            _originalSolutionPath = _thisDir + @$"\{_solutionAndProjectName}";
             _copySolutionPath = _originalSolutionPath + "_";
             _codePath = _copySolutionPath + @"\Executor.cs";
-            PatchSolution(revitVersion);
+            PatchProject(revitVersion);
+            Refresh(false);
         }
 
-        private void PatchSolution(string revitVersion)
+        private void PatchProject(string revitVersion)
         {
-
+            var revitPath = Registry.LocalMachine
+                .OpenSubKey(@$"SOFTWARE\Autodesk\Revit\{revitVersion}\REVIT-05:0419")
+                .GetValue("InstallationLocation");
+            var projectPath = _originalSolutionPath + @$"\{_solutionAndProjectName}.csproj";
+            using var reader = new StreamReader(projectPath);
+            var rewritedProj = string.Format(reader.ReadToEnd(), revitPath);
+            using var writer = new StreamWriter(projectPath);
+            writer.Write(rewritedProj);
         }
 
         public string GetCode()
@@ -38,12 +48,19 @@ namespace RevitRuntimeCompiler.CodeProvider
             return reader.ReadToEnd();
         }
 
-        public void Refresh()
+        public void Refresh() => Refresh(true);
+
+        private void Refresh(bool shouldOverwrite)
         {
             if (Directory.Exists(_copySolutionPath))
+            {
+                if (!shouldOverwrite)
+                    return;
                 Directory.Delete(_copySolutionPath, true);
+            }
             var target = Directory.CreateDirectory(_copySolutionPath);
-            new DirectoryInfo(_originalSolutionPath).CopyFilesRecursively(target);
+            new DirectoryInfo(_originalSolutionPath)
+                .CopyFilesRecursively(target, info => !(info is DirectoryInfo d && d.Name.StartsWith(".")));
         }
     }
 }
